backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
sqlmodel>=0.0.16
psycopg2-binary>=2.9.9
requests>=2.31.0
pandas>=2.2.0
tradingview-ta
pandas-ta
mibian
praw
crewai
langchain-openai
langchain-community
redis
pydantic-settings
